chore: Moved slash command handler
