--- src/main/java/com/greazi/discordbotfoundation/command/SimpleCategory.java
@@ -1,6 +1,0 @@
-package com.greazi.discordbotfoundation.command;
-
-public class SimpleCategory {
-
-
-}
