+++ src/main/java/com/greazi/discordbotfoundation/handlers/commands/SimpleCategory.java
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2021 - 2022. Greazi - All rights reservered
+ *
+ * Unauthorized copying of this file, via any medium is strictly prohibited
+ * Proprietary and confidential
+ */
+
+package com.greazi.discordbotfoundation.handlers.commands;
+
+public class SimpleCategory {
+
+
+}
